+++ eqcorrscan/core/subspace.py
@@ -32,7 +32,7 @@ from obspy.core.event import Event, CreationInfo, ResourceIdentifier, Comment,\
     WaveformStreamID, Pick
 
 from eqcorrscan.utils.clustering import svd
-from eqcorrscan.utils.debug_log import debug_print
+from eqcorrscan.helpers.debug_log import debug_print
 from eqcorrscan.utils import findpeaks, pre_processing, stacking, plotting
 from eqcorrscan.core.match_filter import Detection, extract_from_stream
 from eqcorrscan.utils.plotting import subspace_detector_plot, subspace_fc_plot