+++ eqcorrscan/helpers/__init__.py
@@ -1,0 +1,11 @@
+"""
+:copyright:
+    EQcorrscan developers.
+
+:license:
+    GNU Lesser General Public License, Version 3
+    (https://www.gnu.org/copyleft/lesser.html)
+"""
+if __name__ == '__main__':
+    import doctest
+    doctest.testmod(exclude_empty=True)