+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## Current
+* Added memory management functions in helpers.memory_management
 * Compiled peak-finding routine written to speed-up peak-finding.
 * Add script to test what is the fastest correlation routine for a given dataset
     * This script will write out to a .eqcorrscan.rc file in the users home
@@ -6,6 +7,7 @@
       use a different correlation routine to the standard default.
     * .eqcorrscan.rc can have multiple different defaults for different dataset
       sizes.
+* Moved utils.timer utils.libnames and utils.debug_log to a `helpers` module.
 * Change default match-filter plotting to not decimate unless it has to.
 * BUG-FIX: changed minimum variance for fftw correlation backend.
 * Do not try to process when no processing needs to be done in